+++ src/PowerShell/Commands/DisconnectPartnerCenter.cs
@@ -18,7 +18,6 @@ namespace Microsoft.Store.PartnerCenter.PowerShell.Commands
             if (ShouldProcess(Resources.DisconnectWhatIf))
             {
                 PartnerSession.Instance.Context = null;
-                PartnerService.Instance.EnforceMfa = false;
             }
         }
     }