+++ CHANGELOG.md
@@ -20,6 +20,15 @@
 
 # Change Log
 
+## Upcoming Release
+
+* Authentication
+  * Addressed an issue where NullReferenceException exception was being encountered when invoking [Connect-PartnerCenter](https://docs.microsoft.com/powershell/module/partnercenter/Connect-PartnerCenter) using a certificate
+  * Addressed an issue where NullReferenceException exception was being encountered when invoking [New-PartnerAccessToken](https://docs.microsoft.com/powershell/module/partnercenter/New-PartnerAccessToken) using a certificate
+  * Defined the refresh token parameter set for the [New-PartnerAccessToken](https://docs.microsoft.com/powershell/module/partnercenter/New-PartnerAccessToken) command to make it easier to ensure all the appropriate parameters have been specified when exchanging a refresh token for an access token
+* Module
+  * All commands now perform operations asynchronously
+
 ## 3.0.3 - December 2019
 
 * Authentication